requirements.txt
oracledb~=2.5.1
python-dotenv~=1.0.1
dotenv~=0.9.9
mysql-connector-python~=9.2.0